% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folkeafstemninger.R
\docType{data}
\name{folkeafstemninger}
\alias{folkeafstemninger}
\title{Folkeafstemninger i Danmark}
\format{En data frame med 12 variabler: \code{Dato}, \code{StepType}, \code{StedTekst},
  \code{Partiid}, \code{PartiBogstav}, \code{PartiNavn}, \code{StemmerAntal}, 
  \code{Stemmeberettigede}, \code{DeltagelsePct}, \code{IAltGyldigeStemmer},
  \code{BlankeStemmer} and \code{AndreUgyldigeStemmer}}
\usage{
folkeafstemninger
}
\description{
Et datasæt som indeholder resultaterne af folkeafstemningerne 2009, 2014 og 2015
}
\keyword{datasets}
