% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folkeafstemninger.R
\docType{data}
\name{folkeafstemninger}
\alias{folkeafstemninger}
\title{Folkeafstemninger i Danmark}
\format{En data frame med 9 variabler: \code{Dato}, \code{Landsdel}, \code{Storkreds}, \code{Opstillingskreds},
  \code{Afstemningsomraade}, \code{Parti}, \code{Antal},
  \code{Stemmeberettigede}, \code{DeltagelsePct}}
\usage{
folkeafstemninger
}
\description{
Et datasæt som indeholder resultaterne af folkeafstemningerne 2009, 2014 og 2015
}
\keyword{datasets}
